Major updates to pick a contact

+++ imessage_extractor/src/app/multiapp.py
@@ -16,6 +16,9 @@ from imessage_extractor.src.helpers.utils import ensurelist
 from imessage_extractor.src.app.data.extract import iMessageDataExtract
 
 
+refresh_data = False  # Used for debugging and testing
+
+
 @st.cache(show_spinner=False)
 def save_data_to_local(tmp_env_dpath: str, logger: logging.Logger):
     """
@@ -58,68 +61,76 @@ def save_data_to_local(tmp_env_dpath: str, logger: logging.Logger):
         return fpath
 
 
-    index = ['message_id']
-    fpath = query_and_save_dataset(pg_schema='imessage_extractor',
-                                   dataset_name='message_vw',
-                                   index=index,
-                                   tmp_env_dpath=tmp_env_dpath,
-                                   logger=logger)
+    if refresh_data:
+        index = ['message_id']
+        fpath = query_and_save_dataset(pg_schema='imessage_extractor',
+                                       dataset_name='message_vw',
+                                       index=index,
+                                       tmp_env_dpath=tmp_env_dpath,
+                                       logger=logger)
     manifest['message_vw'] = dict(fpath=fpath, index=index)
 
-    index = ['message_id']
-    fpath = query_and_save_dataset(pg_schema='imessage_extractor',
-                                   dataset_name='message_vw_text',
-                                   index=index,
-                                   tmp_env_dpath=tmp_env_dpath,
-                                   logger=logger)
+    if refresh_data:
+        index = ['message_id']
+        fpath = query_and_save_dataset(pg_schema='imessage_extractor',
+                                       dataset_name='message_vw_text',
+                                       index=index,
+                                       tmp_env_dpath=tmp_env_dpath,
+                                       logger=logger)
     manifest['message_vw_text'] = dict(fpath=fpath, index=index)
 
-    index = ['message_id']
-    fpath = query_and_save_dataset(pg_schema='imessage_extractor',
-                                   dataset_name='message_tokens_unnest',
-                                   index=index,
-                                   tmp_env_dpath=tmp_env_dpath,
-                                   logger=logger)
+    if refresh_data:
+        index = ['message_id']
+        fpath = query_and_save_dataset(pg_schema='imessage_extractor',
+                                       dataset_name='message_tokens_unnest',
+                                       index=index,
+                                       tmp_env_dpath=tmp_env_dpath,
+                                       logger=logger)
     manifest['message_tokens_unnest'] = dict(fpath=fpath, index=index)
 
-    index = ['dt', 'contact_name', 'is_from_me']
-    fpath = query_and_save_dataset(pg_schema='imessage_extractor',
-                                   dataset_name='daily_summary_contact_from_who',
-                                   index=index,
-                                   tmp_env_dpath=tmp_env_dpath,
-                                   logger=logger)
+    if refresh_data:
+        index = ['dt', 'contact_name', 'is_from_me']
+        fpath = query_and_save_dataset(pg_schema='imessage_extractor',
+                                       dataset_name='daily_summary_contact_from_who',
+                                       index=index,
+                                       tmp_env_dpath=tmp_env_dpath,
+                                       logger=logger)
     manifest['daily_summary_contact_from_who'] = dict(fpath=fpath, index=index)
 
-    index = ['dt', 'contact_name']
-    fpath = query_and_save_dataset(pg_schema='imessage_extractor',
-                                   dataset_name='daily_summary_contact',
-                                   index=index,
-                                   tmp_env_dpath=tmp_env_dpath,
-                                   logger=logger)
+    if refresh_data:
+        index = ['dt', 'contact_name']
+        fpath = query_and_save_dataset(pg_schema='imessage_extractor',
+                                       dataset_name='daily_summary_contact',
+                                       index=index,
+                                       tmp_env_dpath=tmp_env_dpath,
+                                       logger=logger)
     manifest['daily_summary_contact'] = dict(fpath=fpath, index=index)
 
-    index = ['dt', 'is_from_me']
-    fpath = query_and_save_dataset(pg_schema='imessage_extractor',
-                                   dataset_name='daily_summary_from_who',
-                                   index=index,
-                                   tmp_env_dpath=tmp_env_dpath,
-                                   logger=logger)
+    if refresh_data:
+        index = ['dt', 'is_from_me']
+        fpath = query_and_save_dataset(pg_schema='imessage_extractor',
+                                       dataset_name='daily_summary_from_who',
+                                       index=index,
+                                       tmp_env_dpath=tmp_env_dpath,
+                                       logger=logger)
     manifest['daily_summary_from_who'] = dict(fpath=fpath, index=index)
 
     index = ['dt']
-    fpath = query_and_save_dataset(pg_schema='imessage_extractor',
-                                   dataset_name='daily_summary',
-                                   index=index,
-                                   tmp_env_dpath=tmp_env_dpath,
-                                   logger=logger)
+    if refresh_data:
+        fpath = query_and_save_dataset(pg_schema='imessage_extractor',
+                                       dataset_name='daily_summary',
+                                       index=index,
+                                       tmp_env_dpath=tmp_env_dpath,
+                                       logger=logger)
     manifest['daily_summary'] = dict(fpath=fpath, index=index)
 
     index = ['contact_name', 'is_from_me', 'token', 'length']
-    fpath = query_and_save_dataset(pg_schema='imessage_extractor',
-                                   dataset_name='contact_token_usage_from_who',
-                                   index=index,
-                                   tmp_env_dpath=tmp_env_dpath,
-                                   logger=logger)
+    if refresh_data:
+        fpath = query_and_save_dataset(pg_schema='imessage_extractor',
+                                       dataset_name='contact_token_usage_from_who',
+                                       index=index,
+                                       tmp_env_dpath=tmp_env_dpath,
+                                       logger=logger)
     manifest['contact_token_usage_from_who'] = dict(fpath=fpath, index=index)
 
     manifest_fpath = join(tmp_env_dpath, 'manifest.json')
@@ -144,14 +155,16 @@ class MultiApp(object):
         self.tmp_env_dpath = join(expanduser('~'), 'Desktop', '.imessage-extractor-app')
 
         # Extract and save data to temporary folder
-        if isdir(self.tmp_env_dpath):
-            rmtree(self.tmp_env_dpath)
+        if refresh_data:
+            if isdir(self.tmp_env_dpath):
+                rmtree(self.tmp_env_dpath)
 
         if not isdir(self.tmp_env_dpath):
             mkdir(self.tmp_env_dpath)
 
-        with st.spinner('Loading iMessage data...'):
-            save_data_to_local(self.tmp_env_dpath, logger)
+        if refresh_data:
+            with st.spinner('Loading iMessage data...'):
+                save_data_to_local(self.tmp_env_dpath, logger)
 
         self.data = iMessageDataExtract(self.tmp_env_dpath, logger)
         self.logger.info('Initialized MultiApp')
